move to f32 for all units, add percentage/absolute widths

+++ src/document.rs
@@ -3,6 +3,6 @@ use rusttype::{point, Font, Scale};
 
 pub struct DocumentDefinition {
     pub font_dict: HashMap<String, Font<'static>>,
-    pub width: u32,
-    pub height: u32,
+    pub width: f32,
+    pub height: f32,
 }
